switched initialization from agent Service to Premain
