+++ rxjava1-finder/src/main/java/com/newrelic/agent/instrumentation/labs/rxjava1/SingleReturnMethodMatcher.java
@@ -1,0 +1,40 @@
+package com.newrelic.agent.instrumentation.labs.rxjava1;
+
+import java.util.Set;
+
+import com.newrelic.agent.deps.org.objectweb.asm.Type;
+import com.newrelic.agent.deps.org.objectweb.asm.commons.Method;
+import com.newrelic.agent.instrumentation.methodmatchers.MethodMatcher;
+import com.newrelic.agent.util.asm.Utils;
+
+public class SingleReturnMethodMatcher implements MethodMatcher {
+	
+	private static final String rxClassname = "rx.Single";
+	
+	public SingleReturnMethodMatcher() {
+	}
+ 
+	@Override
+	public Method[] getExactMethods() {
+		return null;
+	}
+
+	@Override
+	public boolean matches(int access, String name, String desc, Set<String> annotations) {
+		Type type = Type.getReturnType(desc);
+		String classname = type.getClassName();
+		return isSingle(classname);
+	}
+
+	private boolean isSingle(String className) {
+		if(Utils.isPrimitiveType(className) || className.endsWith("[]") || className.startsWith("com.newrelic")) {
+			return false;
+		}
+		if(className.equals(rxClassname)) {
+			return true;
+		}
+		
+		return false;
+	}
+	
+}
